+++ backend/core/src/listings/dto/listing.dto.ts
@@ -1,19 +1,31 @@
 import { Listing } from "../entities/listing.entity"
 import { Expose, Transform, Type } from "class-transformer"
-import { IsDate, IsDefined, IsOptional, IsUUID, ValidateNested } from "class-validator"
+import {
+  ArrayMaxSize,
+  IsDate,
+  IsDefined,
+  IsNumber,
+  IsOptional,
+  IsString,
+  IsUUID,
+  ValidateNested,
+} from "class-validator"
 import moment from "moment"
 import {
   PreferenceCreateDto,
   PreferenceDto,
   PreferenceUpdateDto,
 } from "../../preferences/dto/preference.dto"
-import { OmitType } from "@nestjs/swagger"
+import { ApiProperty, OmitType } from "@nestjs/swagger"
 import { IdDto } from "../../shared/dto/id.dto"
-import { PropertyDto } from "../../property/dto/property.dto"
 import { AddressCreateDto, AddressDto, AddressUpdateDto } from "../../shared/dto/address.dto"
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { UserBasicDto } from "../../user/dto/user.dto"
 import { ListingStatus } from "../types/listing-status-enum"
+import { UnitCreateDto, UnitDto, UnitUpdateDto } from "../../units/dto/unit.dto"
+import { transformUnits } from "../../shared/units-transformations"
+import { UnitsSummarized } from "../../units/types/units-summarized"
+import { Unit } from "../../units/entities/unit.entity"
 
 export class ListingDto extends OmitType(Listing, [
   "preferences",
@@ -30,12 +42,6 @@ export class ListingDto extends OmitType(Listing, [
   @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
   @Type(() => PreferenceDto)
   preferences: PreferenceDto[]
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => PropertyDto)
-  property: PropertyDto
 
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
@@ -71,6 +77,179 @@ export class ListingDto extends OmitType(Listing, [
     return listing.status
   })
   status: ListingStatus
+
+  @Expose()
+  @Type(() => UnitDto)
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.units
+    },
+    { toClassOnly: true }
+  )
+  units: UnitDto[]
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.accessibility
+    },
+    { toClassOnly: true }
+  )
+  accessibility: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.amenities
+    },
+    { toClassOnly: true }
+  )
+  amenities: string | null
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AddressDto)
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.buildingAddress
+    },
+    { toClassOnly: true }
+  )
+  buildingAddress: AddressDto
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.buildingTotalUnits
+    },
+    { toClassOnly: true }
+  )
+  buildingTotalUnits: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.developer
+    },
+    { toClassOnly: true }
+  )
+  developer: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.householdSizeMax
+    },
+    { toClassOnly: true }
+  )
+  householdSizeMax: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.householdSizeMin
+    },
+    { toClassOnly: true }
+  )
+  householdSizeMin: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.neighborhood
+    },
+    { toClassOnly: true }
+  )
+  neighborhood: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.petPolicy
+    },
+    { toClassOnly: true }
+  )
+  petPolicy: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.smokingPolicy
+    },
+    { toClassOnly: true }
+  )
+  smokingPolicy: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.unitsAvailable
+    },
+    { toClassOnly: true }
+  )
+  unitsAvailable: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.unitAmenities
+    },
+    { toClassOnly: true }
+  )
+  unitAmenities: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.servicesOffered
+    },
+    { toClassOnly: true }
+  )
+  servicesOffered?: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value, obj: Listing) => {
+      return obj.property.yearBuilt
+    },
+    { toClassOnly: true }
+  )
+  yearBuilt: number | null
+
+  @Expose()
+  @ApiProperty({ type: UnitsSummarized })
+  get unitsSummarized(): UnitsSummarized | undefined {
+    if (Array.isArray(this.units) && this.units.length > 0) {
+      return transformUnits(this.units as Unit[])
+    }
+  }
 }
 
 export class ListingCreateDto extends OmitType(ListingDto, [
@@ -78,25 +257,34 @@ export class ListingCreateDto extends OmitType(ListingDto, [
   "createdAt",
   "updatedAt",
   "preferences",
-  "property",
   "applicationAddress",
   "applicationPickUpAddress",
   "leasingAgentAddress",
   "leasingAgents",
   "urlSlug",
   "showWaitlist",
+  "units",
+  "accessibility",
+  "amenities",
+  "buildingAddress",
+  "buildingTotalUnits",
+  "developer",
+  "householdSizeMax",
+  "householdSizeMin",
+  "neighborhood",
+  "petPolicy",
+  "smokingPolicy",
+  "unitsAvailable",
+  "unitAmenities",
+  "servicesOffered",
+  "yearBuilt",
+  "unitsSummarized",
 ] as const) {
   @Expose()
   @IsDefined({ groups: [ValidationsGroupsEnum.default] })
   @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
   @Type(() => PreferenceCreateDto)
   preferences: PreferenceCreateDto[]
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  property: IdDto
 
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
@@ -122,6 +310,84 @@ export class ListingCreateDto extends OmitType(ListingDto, [
   @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
   @Type(() => IdDto)
   leasingAgents?: IdDto[] | null
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
+  @ArrayMaxSize(256, { groups: [ValidationsGroupsEnum.default] })
+  @Type(() => UnitCreateDto)
+  units: UnitCreateDto[]
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  accessibility: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  amenities: string | null
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AddressCreateDto)
+  buildingAddress: AddressCreateDto
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  buildingTotalUnits: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  developer: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  householdSizeMax: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  householdSizeMin: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  neighborhood: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  petPolicy: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  smokingPolicy: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  unitsAvailable: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  unitAmenities: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  servicesOffered?: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  yearBuilt: number | null
 }
 
 export class ListingUpdateDto extends OmitType(ListingDto, [
@@ -129,13 +395,28 @@ export class ListingUpdateDto extends OmitType(ListingDto, [
   "createdAt",
   "updatedAt",
   "preferences",
-  "property",
   "applicationAddress",
   "applicationPickUpAddress",
   "leasingAgentAddress",
   "urlSlug",
   "leasingAgents",
   "showWaitlist",
+  "units",
+  "accessibility",
+  "amenities",
+  "buildingAddress",
+  "buildingTotalUnits",
+  "developer",
+  "householdSizeMax",
+  "householdSizeMin",
+  "neighborhood",
+  "petPolicy",
+  "smokingPolicy",
+  "unitsAvailable",
+  "unitAmenities",
+  "servicesOffered",
+  "yearBuilt",
+  "unitsSummarized",
 ] as const) {
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
@@ -161,12 +442,6 @@ export class ListingUpdateDto extends OmitType(ListingDto, [
   preferences: PreferenceUpdateDto[]
 
   @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  property: IdDto
-
-  @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
   @Type(() => AddressUpdateDto)
@@ -190,4 +465,82 @@ export class ListingUpdateDto extends OmitType(ListingDto, [
   @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
   @Type(() => IdDto)
   leasingAgents?: IdDto[] | null
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
+  @ArrayMaxSize(256, { groups: [ValidationsGroupsEnum.default] })
+  @Type(() => UnitUpdateDto)
+  units: UnitUpdateDto[]
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  accessibility: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  amenities: string | null
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AddressUpdateDto)
+  buildingAddress: AddressUpdateDto
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  buildingTotalUnits: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  developer: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  householdSizeMax: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  householdSizeMin: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  neighborhood: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  petPolicy: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  smokingPolicy: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  unitsAvailable: number | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  unitAmenities: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  servicesOffered?: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
+  yearBuilt: number | null
 }