+++ backend/core/test/listings/listings.e2e-spec.ts
@@ -4,6 +4,9 @@ import { ListingsModule } from "../../src/listings/listings.module"
 import supertest from "supertest"
 import { applicationSetup } from "../../src/app.module"
 import { allSeeds } from "../../src/seeds/listings"
+import { ListingDto } from "../../src/listings/dto/listing.dto"
+import { getUserAccessToken } from "../utils/get-user-access-token"
+import { setAuthorization } from "../utils/set-authorization-helper"
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const dbOptions = require("../../ormconfig.test")
@@ -48,6 +51,31 @@ describe("Listings", () => {
     const query = "/?jsonpath=%24%5B%3F%28%40.status%3D%3D%22active%22%29%5D"
     const res = await supertest(app.getHttpServer()).get(`/listings${query}`).expect(200)
     expect(res.body.length).toEqual(allSeeds.length)
+  })
+
+  it("should modify property related fields of a listing and return a modified value", async () => {
+    const res = await supertest(app.getHttpServer()).get("/listings").expect(200)
+
+    const listing: ListingDto = { ...res.body[0] }
+
+    const amenitiesValue = "Random amenities value"
+    expect(listing.amenities).not.toBe(amenitiesValue)
+    listing.amenities = amenitiesValue
+
+    const oldOccupancy = listing.units[0].maxOccupancy
+    listing.units[0].maxOccupancy = oldOccupancy + 1
+
+    const adminAccessToken = await getUserAccessToken(app, "admin@example.com", "abcdef")
+
+    const putResponse = await supertest(app.getHttpServer())
+      .put(`/listings/${listing.id}`)
+      .send(listing)
+      .set(...setAuthorization(adminAccessToken))
+      .expect(200)
+    const modifiedListing: ListingDto = putResponse.body
+
+    expect(modifiedListing.amenities).toBe(amenitiesValue)
+    expect(modifiedListing.units[0].maxOccupancy).toBe(oldOccupancy + 1)
   })
 
   afterEach(() => {