+++ ui-components/__tests__/page_components/Waitlist.test.tsx
@@ -3,19 +3,21 @@ import { render, cleanup } from "@testing-library/react"
 import { Waitlist } from "../../src/page_components/listing/listing_sidebar/Waitlist"
 import Archer from "../fixtures/archer.json"
 import Triton from "../fixtures/triton-test.json"
+import { Listing } from "@bloom-housing/backend-core/types"
 
-const archer = Object.assign({}, Archer) as any
-const triton = Object.assign({}, Triton) as any
-archer.property = {}
-archer.property.unitsSummarized = {}
-archer.property.unitsSummarized.byNonReservedUnitType = []
-archer.property.unitsSummarized.byReservedType = []
+const archer: Listing = Object.assign({}, Archer) as any
+const triton: Listing = Object.assign({}, Triton) as any
+
+// @ts-ignore
+archer.unitsSummarized = {}
+archer.unitsSummarized.byNonReservedUnitType = []
+archer.unitsSummarized.byReservedType = []
 archer.waitlistCurrentSize = 300
 
-triton.property = {}
-triton.property.unitsSummarized = {}
-triton.property.unitsSummarized.byNonReservedUnitType = []
-triton.property.unitsSummarized.byReservedType = []
+// @ts-ignore
+triton.unitsSummarized = {}
+triton.unitsSummarized.byNonReservedUnitType = []
+triton.unitsSummarized.byReservedType = []
 
 afterEach(cleanup)
 
@@ -27,7 +29,10 @@ describe("<Waitlist>", () => {
     expect(getByText("Final Waitlist Size"))
     expect(getAllByText("300").length).toBe(2)
   })
-  it("renders with and open waitlist", () => {
+  it("renders with an open waitlist", () => {
+    triton.unitsAvailable = 0
+    triton.waitlistCurrentSize = 40
+    triton.waitlistMaxSize = 100
     const { getByText } = render(<Waitlist listing={triton} />)
     expect(getByText("Waitlist is open")).toBeTruthy()
     expect(getByText("Current Waitlist Size"))
@@ -38,7 +43,7 @@ describe("<Waitlist>", () => {
   })
   it("renders with open spots", () => {
     const newListing = triton
-    newListing.property.unitsAvailable = 1
+    newListing.unitsAvailable = 1
     newListing.waitlistCurrentSize = 0
     newListing.waitlistMaxSize = 10
     const { getByText } = render(<Waitlist listing={newListing} />)